Add Maps.put_by_id helper
