+++ lib/chukinas/util/precision.ex
@@ -33,12 +33,4 @@ defmodule Chukinas.Util.Precision do
   def approx_equal(a, b) do
     abs(a - b) < 1
   end
-
-  # TODO move this to a better-named module
-  # TODO use this in the several places in sprite.ex
-  def apply_to_each_val(%{} = map, fun) do
-    map
-    |> Stream.map(fn {key, val} -> {key, fun.(val)} end)
-    |> Enum.into(%{})
-  end
 end
